Add new typings
